requirements.txt
# Coffee Shop AI Agent - Python 3.13 Compatible Requirements

# FastAPI and ASGI server
fastapi==0.115.6
uvicorn[standard]==0.33.0

# WebSocket support
websockets==14.0

# Redis for session management
redis==5.2.1
hiredis==3.1.0

# HTTP client for API calls
httpx==0.28.1
aiohttp==3.11.9

# Environment variables
python-dotenv==1.0.1

# Data validation (Python 3.13 compatible)
pydantic==2.10.4

# JSON handling
orjson==3.10.11

# Logging
structlog==24.4.0

# Google AI (Gemini) - Updated for Python 3.13
google-generativeai==0.8.3

# Date/time utilities
python-dateutil==2.9.0.post0

# Static file serving
aiofiles==24.1.0

# CORS middleware
python-multipart==0.0.19

# Payment Processing
stripe==11.1.0

# Email notifications
emails==0.6

# QR Code generation (for virtual queue)
qrcode[pil]==7.4.2

# UUID utilities
shortuuid==1.0.13

# Development dependencies
pytest==8.3.4
pytest-asyncio==0.24.0

fastapi==0.115.6
uvicorn[standard]==0.33.0
websockets==14.0
redis==5.2.1
hiredis==3.1.0
httpx==0.28.1
aiohttp==3.11.9
python-dotenv==1.0.1
pydantic==2.10.4
orjson==3.10.11
structlog==24.4.0
google-generativeai==0.8.3
python-dateutil==2.9.0.post0
aiofiles==24.1.0
python-multipart==0.0.19
stripe==11.1.0
emails==0.6
qrcode[pil]==7.4.2
shortuuid==1.0.13



